+++ src/Kook.Net.Rest/API/Rest/ModifyDirectMessageParams.cs
@@ -10,7 +10,7 @@ internal class ModifyDirectMessageParams
 
     [JsonPropertyName("template_id")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? TemplateId { get; set; }
+    public ulong? TemplateId { get; set; }
 
     [JsonPropertyName("content")]
     public required string Content { get; set; }
