+++ test/Kook.Net.Tests.Unit/MockedEntities/MockedDMChannel.cs
@@ -42,7 +42,7 @@ internal sealed class MockedDMChannel : IDMChannel
         throw new NotImplementedException();
 
     /// <inheritdoc />
-    public Task<Cacheable<IUserMessage, Guid>> SendTextAsync<T>(int templateId, T parameters, IQuote? quote = null,
+    public Task<Cacheable<IUserMessage, Guid>> SendTextAsync<T>(ulong templateId, T parameters, IQuote? quote = null,
         JsonSerializerOptions? jsonSerializerOptions = null, RequestOptions? options = null) =>
         throw new NotImplementedException();
 
@@ -57,7 +57,7 @@ internal sealed class MockedDMChannel : IDMChannel
         throw new NotImplementedException();
 
     /// <inheritdoc />
-    public Task<Cacheable<IUserMessage, Guid>> SendCardsAsync<T>(int templateId, T parameters, IQuote? quote = null,
+    public Task<Cacheable<IUserMessage, Guid>> SendCardsAsync<T>(ulong templateId, T parameters, IQuote? quote = null,
         JsonSerializerOptions? jsonSerializerOptions = null, RequestOptions? options = null) =>
         throw new NotImplementedException();
 
@@ -102,7 +102,7 @@ internal sealed class MockedDMChannel : IDMChannel
         throw new NotImplementedException();
 
     /// <inheritdoc />
-    public Task<Cacheable<IUserMessage, Guid>> SendTextAsync<T>(int templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
+    public Task<Cacheable<IUserMessage, Guid>> SendTextAsync<T>(ulong templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
         JsonSerializerOptions? jsonSerializerOptions = null, RequestOptions? options = null) =>
         throw new NotImplementedException();
 
@@ -117,7 +117,7 @@ internal sealed class MockedDMChannel : IDMChannel
         throw new NotImplementedException();
 
     /// <inheritdoc />
-    public Task<Cacheable<IUserMessage, Guid>> SendCardsAsync<T>(int templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
+    public Task<Cacheable<IUserMessage, Guid>> SendCardsAsync<T>(ulong templateId, T parameters, IQuote? quote = null, IUser? ephemeralUser = null,
         JsonSerializerOptions? jsonSerializerOptions = null, RequestOptions? options = null) =>
         throw new NotImplementedException();
